Refactor command handlers: consolidate create, grant, and update functionalities into commands package; update file paths for database access
