+++ commands/create/create_test.go
@@ -1,4 +1,4 @@
-package create_test
+package commands_test
 
 import (
 	"testing"
