+++ commands/create/sercive.go
@@ -1,4 +1,4 @@
-package create
+package commands
 
 import (
